refactor(config): extract tools configuration module
